Optimize PreferenceColor
refactor: Extract ColorPicker into a standalone component

+++ composeApp/src/commonMain/kotlin/xyz/hyli/timeflow/datastore/DataStore.kt
@@ -59,7 +59,7 @@ class SettingsDataStore(
             currentSettings.copy(themeDynamicColor = themeDynamicColor)
         }
     }
-    suspend fun updateThemeColor(color: Long) {
+    suspend fun updateThemeColor(color: Int) {
         db.updateData { currentSettings ->
             currentSettings.copy(themeColor = color)
         }
